python/requirements.txt
pycoingecko == 3.1.0
binance-connector
pandas == 1.5.1
SQLAlchemy == 1.4.44

